Remove shared state (AddressingProperties) taking WinRmClient closer to thread safety
AddressingProperties was shared between all requests, making it impossible to use the same WinRmClient concurrently. It stores things like the current action name (which we had to update because it takes the first one otherwise), the message id. We needed that so we can force CXF into using the 2004 version of the addressing namespace. The changes here take a different approach (creating a policy and setting a special assertion on it), causing CXF to create the AddressingProperties object per request.